refactor: OverlaidPopup 컴포넌트 제작
