+++ front/src/Components/Custom/TrimOptionGroup/index.jsx
@@ -4,22 +4,27 @@ import Button from "../../Common/Button/Button";
 import TrimOption from "../TrimOption";
 import Tooltip from "../Tooltip";
 import useTooltip from "../../../hooks/useTooltip";
-import BackgroundOverlay from "../../Common/BackgroundOverlay";
 import TrimComparisonPopup from "../../../Pages/CustomPage/TrimPage/TrimComparisonPopup";
 import useOnClickPopUp from "../../../hooks/useOnClickPopUp";
+import OverlaidPopup from "../../Common/OverlaidPopup";
 
 const TrimOptionGroup = ({ options, selectedOption, handleOptionSelect }) => {
   const { isTooltipOpen, openTooltip, closeTooltip } = useTooltip();
   const popupRef = useRef();
-  const { isPopupOpen, openPopup, closePopup } = useOnClickPopUp(popupRef);
+  const {
+    isPopupOpen: isComparisonPopupOpen,
+    openPopup,
+    closePopup,
+  } = useOnClickPopUp(popupRef);
   return (
     <Wrapper>
       {isTooltipOpen && <Tooltip offset={810} />}
-      {isPopupOpen && (
-        <PopupWrapper>
-          <BackgroundOverlay />
-          <TrimComparisonPopup popupRef={popupRef} closePopup={closePopup} />
-        </PopupWrapper>
+      {isComparisonPopupOpen && (
+        <OverlaidPopup
+          component={
+            <TrimComparisonPopup popupRef={popupRef} closePopup={closePopup} />
+          }
+        />
       )}
       <Title onMouseEnter={openTooltip} onMouseLeave={closeTooltip}>
         <span>트림</span>
